requirement.txt
crewai
crewai_tools
load_dotenv
langchain
langchain-huggingface
tavily-python
pandas